backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
python-jose
python-multipart
pydantic
requests
email-validator
openai
python-docx
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
httplib2
oauth2client
lxml
